v1.2.1 1.实现折扣系统2.网格布局也改成像输出那个按钮一样记录状态吧，这样玩家打开商店都可以默认使用自己习惯的布局3.网格布局把买/卖的字放到数量的左侧4.ftb的按钮打开添加一个配置项来选择默认打开的商店，如果不配置该配置项则正常打开选择UI5.指令这一块改成List<String> 不需要自己用;分割多个指令6.修复int导致的bug7.打开商店的那个ui也支持一下多主题8.优化编辑器组件，使得编辑起来更加方便

+++ src/main/java/com/viscriptshop/gui/data/ItemMatchRule.java
@@ -48,11 +48,11 @@ public class ItemMatchRule implements IConfigurable, IPersistedSerializable {
         return ItemUtil.isSameItem(candidate, target, resolvedCompareMode(), resolvedComponents());
     }
 
-    public int getItemForPlayerCount(ServerPlayer player, ItemStack itemStack) {
+    public long getItemForPlayerCount(ServerPlayer player, ItemStack itemStack) {
         return ItemUtil.getItemForPlayerCount(player, itemStack, resolvedCompareMode(), resolvedComponents());
     }
 
-    public void removeItemForPlayer(ServerPlayer player, ItemStack itemStack, int count) {
+    public void removeItemForPlayer(ServerPlayer player, ItemStack itemStack, long count) {
         ItemUtil.removeItemForPlayer(player, itemStack, count, resolvedCompareMode(), resolvedComponents());
     }
